refactoring the function structure.
